Make get_grid_row_count count actual rows
Also make it a bit smarter. Specifically, rather than looking at the
distance from the line to the origin (which can be misleading for
slanted lines), look at the row coordinate of its horizontally-centered
point. Also restrict the Fourier transform to the actual range of
observed line points, so that we're not including extra space at the top
and bottom.